rename _conditional_bernoulli to _combinatorics
