+++ src/pydrobert/torch/_combinatorics.py
