+++ tests/test_combinatorics.py
